Write SSH host key to console
This fixes an issue with the Cloudscale Terraform provider. The provider
failed to detect when the server was ready. This was caused by the full
SSH host key no longer was written to console.

Cloudscale will provide an option to ignore this. Meanwhile this has to
be used as a workaround.